Typecasts the hash as an unsigned integer.

+++ Contract.sol
@@ -17,7 +17,7 @@ contract ZombieFactory {
     } //end function createZombie()
 
     function _generateRandomDna(string memory _str) private view returns (uint) {
-        uint rand = keccak256(abi.encodePacked(_str));
+        uint rand = uint(keccak256(abi.encodePacked(_str)));
         return rand % dnaModulus;
     } //end function _generateRandomDna()
 } //end contract ZombieFactory {}
